Refactor UnifiedAllocator to take in the kernel arch
This change is to decouple the allocator from just x86/cuda archs. As
part of the refactor, I also moved the Arch enum into its own file.
